CSkeleton: Make NumBones() return a size_t
Consistent with the rest of the API.

+++ src/Core/Resource/Animation/CSkeleton.h
@@ -43,7 +43,7 @@ public:
     void Draw(FRenderOptions Options, const CBoneTransformData *pkData);
     std::pair<int32, float> RayIntersect(const CRay& rkRay, const CBoneTransformData& rkData) const;
 
-    uint32 NumBones() const  { return mBones.size(); }
+    size_t NumBones() const  { return mBones.size(); }
     CBone* RootBone() const  { return mpRootBone; }
 };
